Fix SonarQube issues: split large files, extract functions
Split helpers.js into helpers.js + helpers-prn.js (PRN transaction
builders and performers). Split update-status.js by extracting balance
effect functions into update-status-balance-effects.js. Extract
appendEvent logic from stream-inmemory.js factory to standalone
function. Use Set for PRN_KINDS and .at(-1) in stream files.
